+++ src/main/java/io/dataspaceconnector/common/net/HttpService.java
@@ -12,6 +12,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ *  Contributors:
+ *       sovity GmbH
  */
 package io.dataspaceconnector.common.net;
 
@@ -28,6 +31,7 @@ import okhttp3.HttpUrl;
 import okhttp3.MediaType;
 import okhttp3.Request;
 import okhttp3.RequestBody;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
 
 import java.io.ByteArrayInputStream;
@@ -117,8 +121,7 @@ public class HttpService implements DataRetrievalService {
      */
     public Response post(final URL target, final HttpArgs args, final InputStream data)
             throws IOException {
-        Utils.requireNonNull(target, ErrorMessage.URI_NULL);
-        Utils.requireNonNull(args, ErrorMessage.HTTP_ARGS_NULL);
+        validateParameter(target, args);
 
         final var urlBuilder = createUrlBuilder(target);
 
@@ -140,15 +143,16 @@ public class HttpService implements DataRetrievalService {
         }
 
         final var response = httpSvc.send(requestBuilder.build());
-
-        final var output = new HttpResponse(response.code(), getBody(response));
+        final var output = getHttpResponse(response);
         response.close();
 
         return output;
     }
 
     /**
-     * Perform a get request.
+     * Perform a get request with query parameters (if given).
+     * If a query parameter is already found in the target URL it is ignored,
+     * so that the basis url of the target can not be bypassed.
      *
      * @param target The recipient of the request.
      * @param args   The request arguments.
@@ -157,14 +161,17 @@ public class HttpService implements DataRetrievalService {
      * @throws IllegalArgumentException if any of the parameters is null.
      */
     public Response get(final URL target, final HttpArgs args) throws IOException {
-        Utils.requireNonNull(target, ErrorMessage.URI_NULL);
-        Utils.requireNonNull(args, ErrorMessage.HTTP_ARGS_NULL);
+        validateParameter(target, args);
 
         final var urlBuilder = createUrlBuilder(target);
+        final var targetUrl = toUrl(target);
+        final var queryKeys = targetUrl.queryParameterNames();
 
         if (args.getParams() != null) {
             for (final var key : args.getParams().keySet()) {
-                urlBuilder.addQueryParameter(key, args.getParams().get(key));
+                if (!queryKeys.contains(key)) {
+                    urlBuilder.addQueryParameter(key, args.getParams().get(key));
+                }
             }
         }
 
@@ -186,10 +193,20 @@ public class HttpService implements DataRetrievalService {
             response = httpSvc.getWithHeaders(targetUri, headerCopy);
         }
 
-        final var output = new HttpResponse(response.code(), getBody(response));
+        final var output = getHttpResponse(response);
         response.close();
 
         return output;
+    }
+
+    @NotNull
+    private HttpResponse getHttpResponse(final okhttp3.Response response) throws IOException {
+        return new HttpResponse(response.code(), getBody(response));
+    }
+
+    private void validateParameter(final URL target, final HttpArgs args) {
+        Utils.requireNonNull(target, ErrorMessage.URI_NULL);
+        Utils.requireNonNull(args, ErrorMessage.HTTP_ARGS_NULL);
     }
 
     private InputStream getBody(final okhttp3.Response response) throws IOException {