+++ src/test/java/io/dataspaceconnector/common/net/HttpServiceTest.java
@@ -12,6 +12,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ *  Contributors:
+ *       sovity GmbH
  */
 package io.dataspaceconnector.common.net;
 
@@ -228,5 +231,41 @@ class HttpServiceTest {
         assertNotNull(result);
         assertEquals(responseCode, result.getCode());
         assertArrayEquals(bytes, result.getData().readAllBytes());
+    }
+
+    @Test
+    @SneakyThrows
+    void get_withAlreadyDefinedQuery() {
+        /* ARRANGE */
+        final var params = Map.of("name", "doe");
+        final var args = new HttpService.HttpArgs();
+        args.setParams(params);
+
+        final var target = new URL("https://target.com/?name=john");
+
+        final var response = new Response.Builder()
+                .request(new Request.Builder().url(target).build())
+                .protocol(Protocol.HTTP_1_1).code(200).message("Some message")
+                .body(ResponseBody.create("someBody", MediaType.parse("application/text")))
+                .build();
+
+        // The first response will be consumed by the first request, duplicate
+        final var response2 = new Response.Builder()
+                .request(new Request.Builder().url(target).build())
+                .protocol(Protocol.HTTP_1_1).code(200).message("Some message")
+                .body(ResponseBody.create("someBody", MediaType.parse("application/text")))
+                .build();
+
+        Mockito.doReturn(response).when(httpSvc).get(any());
+
+        /* ACT */
+        final var result = (HttpResponse) service.request(HttpService.Method.GET, target, args);
+
+        /* ASSERT */
+        Mockito.doReturn(response2).when(httpSvc).get(any());
+        final var expected = (HttpResponse) service.get(target, args);
+        assertEquals(expected.getCode(), result.getCode());
+        assertTrue(Arrays.areEqual("someBody".getBytes(StandardCharsets.UTF_8),
+                result.getData().readAllBytes()));
     }
 }